feat : 마무리 수정

+++ NADRI-GIL/src/main/java/back/NADRIGIL/dto/travel/GetRecommendTravelListDTO.java
@@ -10,5 +10,5 @@ public class GetRecommendTravelListDTO {
     private String name;
     private String location;
     private String image;
-    private Long similarity;
+    private int similarity;
 }
